[Fix #634] Refactoring modules
