+++ api/pom.xml
@@ -14,29 +14,23 @@
 
     <dependencies>
         <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
+            <groupId>io.serverlessworkflow</groupId>
+            <artifactId>serverlessworkflow-types</artifactId>
+            <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-core</artifactId>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
         </dependency>
         <dependency>
           <groupId>com.networknt</groupId>
           <artifactId>json-schema-validator</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-        </dependency>
-        <dependency>
             <groupId>com.fasterxml.jackson.dataformat</groupId>
             <artifactId>jackson-dataformat-yaml</artifactId>
         </dependency>
-        <dependency>
-            <groupId>jakarta.validation</groupId>
-            <artifactId>jakarta.validation-api</artifactId>
-        </dependency>
+        
         <dependency>
           <groupId>org.hibernate.validator</groupId>
           <artifactId>hibernate-validator</artifactId>
@@ -79,54 +73,4 @@
             <scope>test</scope>
         </dependency>
     </dependencies>
-
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.jsonschema2pojo</groupId>
-                <artifactId>jsonschema2pojo-maven-plugin</artifactId>
-                <configuration>
-                   <sourceDirectory>${basedir}/src/main/resources/schema</sourceDirectory>
-                   <!--The comment below is left intentionally in case jsonschema2pojo one day accepts https urls. That day we can remove the file from schema dir and use directly the real schema-->
-                   <!-- <sourcePaths>
-                      <sourcePath>https://raw.githubusercontent.com/serverlessworkflow/specification/main/schema/workflow.yaml</sourcePath>
-                    </sourcePaths> -->
-                    <sourceType>yamlschema</sourceType>
-                    <targetPackage>io.serverlessworkflow.api.types</targetPackage>
-                    <outputDirectory>${project.build.directory}/generated-sources/src/main/java</outputDirectory>
-                    <includeJsr303Annotations>true</includeJsr303Annotations>
-                    <generateBuilders>true</generateBuilders>
-                    <initializeCollections>true</initializeCollections>
-                    <includeAdditionalProperties>true</includeAdditionalProperties>
-                    <includeToString>false</includeToString>
-                    <includeHashcodeAndEquals>false</includeHashcodeAndEquals>
-                    <includeConstructors>true</includeConstructors>
-                    <constructorsRequiredPropertiesOnly>true</constructorsRequiredPropertiesOnly>
-                    <useTitleAsClassname>true</useTitleAsClassname>
-                    <serializable>true</serializable>
-                    <targetVersion>${java.version}</targetVersion>
-                    <usePrimitives>true</usePrimitives>
-                    <useJakartaValidation>true</useJakartaValidation>
-                    <customRuleFactory>io.serverlessworkflow.generator.UnreferencedFactory</customRuleFactory>
-                    <customAnnotator>io.serverlessworkflow.generator.ConstAnnotator</customAnnotator>
-                </configuration>
-                <dependencies>
-                    <dependency>
-                        <groupId>io.serverlessworkflow</groupId>
-                        <artifactId>serverless-workflow-custom-generator</artifactId>
-                        <version>${project.version}</version>
-                    </dependency>
-                </dependencies>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>generate</goal>
-                        </goals>
-                        <phase>generate-sources</phase>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-</project>
+ </project>
