+++ custom-generator/src/main/java/io/serverlessworkflow/generator/AllAnyOneOfSchemaRule.java
@@ -33,6 +33,10 @@ import com.sun.codemodel.JMod;
 import com.sun.codemodel.JPackage;
 import com.sun.codemodel.JType;
 import com.sun.codemodel.JVar;
+import io.serverlessworkflow.annotations.OneOfSetter;
+import io.serverlessworkflow.annotations.OneOfValueProvider;
+import io.serverlessworkflow.serialization.DeserializeHelper;
+import io.serverlessworkflow.serialization.SerializeHelper;
 import jakarta.validation.ConstraintViolationException;
 import java.io.UnsupportedEncodingException;
 import java.net.URI;
@@ -318,10 +322,7 @@ class AllAnyOneOfSchemaRule extends SchemaRule {
             null);
 
     definedClass._implements(
-        definedClass
-            .owner()
-            .ref(GeneratorUtils.ONE_OF_VALUE_PROVIDER_INTERFACE_NAME)
-            .narrow(valueField.type()));
+        definedClass.owner().ref(OneOfValueProvider.class).narrow(valueField.type()));
     GeneratorUtils.implementInterface(definedClass, valueField);
     try {
       JDefinedClass serializer = generateSerializer(definedClass);
@@ -397,9 +398,7 @@ class AllAnyOneOfSchemaRule extends SchemaRule {
         (method, valueParam, genParam) ->
             method
                 .body()
-                .staticInvoke(
-                    definedClass.owner().ref(GeneratorUtils.SERIALIZE_HELPER_NAME),
-                    "serializeOneOf")
+                .staticInvoke(definedClass.owner().ref(SerializeHelper.class), "serializeOneOf")
                 .arg(genParam)
                 .arg(valueParam));
     return definedClass;
@@ -418,7 +417,7 @@ class AllAnyOneOfSchemaRule extends SchemaRule {
           body._return(
               definedClass
                   .owner()
-                  .ref(GeneratorUtils.DESERIALIZE_HELPER_NAME)
+                  .ref(DeserializeHelper.class)
                   .staticInvoke(methodName)
                   .arg(parserParam)
                   .arg(relatedClass.dotclass())
@@ -460,7 +459,7 @@ class AllAnyOneOfSchemaRule extends SchemaRule {
         v -> {
           method.body().assign(JExpr._this().ref(v), methodParam);
           method
-              .annotate(definedClass.owner().ref(GeneratorUtils.SETTER_ANNOTATION_NAME))
+              .annotate(definedClass.owner().ref(OneOfSetter.class))
               .param("value", instanceField.type());
         });
     return methodParam;
