+++ serverlessworkflow-annotations/src/main/java/io/serverlessworkflow/annotations/OneOfValueProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.serverlessworkflow.api;
+package io.serverlessworkflow.annotations;
 
 public interface OneOfValueProvider<T> {
   T get();
