+++ serverlessworkflow-types/src/main/resources/schema/workflow.yaml
